requirements.txt
pyyaml
docker-compose